# swagger_latest.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SchoolERPManagementAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Exams": {
      "post": {
        "tags": [
          "Exams"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExamDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Exams/publish-result": {
      "post": {
        "tags": [
          "Exams"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishResultDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishResultDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PublishResultDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Exams/student/{studentId}/results": {
      "get": {
        "tags": [
          "Exams"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Fees/pay": {
      "post": {
        "tags": [
          "Fees"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeePaymentDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FeePaymentDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FeePaymentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Fees/student/{studentId}/summary": {
      "get": {
        "tags": [
          "Fees"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Fees/student/{studentId}/history": {
      "get": {
        "tags": [
          "Fees"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Fees/create-checkout-session": {
      "post": {
        "tags": [
          "Fees"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckoutSessionDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Homework": {
      "post": {
        "tags": [
          "Homework"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "SubjectId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "TeacherId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "ClassId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "Title": {
                    "type": "string"
                  },
                  "Description": {
                    "type": "string"
                  },
                  "Attachment": {
                    "type": "string",
                    "format": "binary"
                  },
                  "DueDate": {
                    "type": "string",
                    "format": "date"
                  }
                }
              },
              "encoding": {
                "SubjectId": {
                  "style": "form"
                },
                "TeacherId": {
                  "style": "form"
                },
                "ClassId": {
                  "style": "form"
                },
                "Title": {
                  "style": "form"
                },
                "Description": {
                  "style": "form"
                },
                "Attachment": {
                  "style": "form"
                },
                "DueDate": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Homework/submit": {
      "post": {
        "tags": [
          "Homework"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "HomeworkId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "StudentId": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "UploadedFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "HomeworkId": {
                  "style": "form"
                },
                "StudentId": {
                  "style": "form"
                },
                "UploadedFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Homework/evaluate": {
      "post": {
        "tags": [
          "Homework"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateHomeworkDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateHomeworkDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateHomeworkDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Notifications": {
      "post": {
        "tags": [
          "Notifications"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Notifications/user/{userId}": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Parents/{id}": {
      "get": {
        "tags": [
          "Parents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Parents/{id}/children": {
      "get": {
        "tags": [
          "Parents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Parents": {
      "post": {
        "tags": [
          "Parents"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParentDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParentDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParentDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCheckoutSessionDTO": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "successUrl": {
            "type": "string",
            "nullable": true
          },
          "cancelUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CreateExamDTO": {
        "type": "object",
        "properties": {
          "examname": {
            "type": "string",
            "nullable": true
          },
          "academicyearId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateParentDTO": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "relation": {
            "type": "string",
            "nullable": true
          },
          "phonenumber": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "EvaluateHomeworkDTO": {
        "type": "object",
        "properties": {
          "homeworkSubmissionId": {
            "type": "integer",
            "format": "int32"
          },
          "marks": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "remarks": {
            "type": "string",
            "nullable": true
          },
          "verificationStatus": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FeePaymentDTO": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "amountPaid": {
            "type": "number",
            "format": "double"
          },
          "paymentDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "paymentMethod": {
            "type": "string",
            "nullable": true
          },
          "transactionId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PublishResultDTO": {
        "type": "object",
        "properties": {
          "examId": {
            "type": "integer",
            "format": "int32"
          },
          "subjectId": {
            "type": "integer",
            "format": "int32"
          },
          "studentId": {
            "type": "integer",
            "format": "int32"
          },
          "marks": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "uploadedCorrectedAnswerSheetUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SendNotificationDTO": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "createdByUserId": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "targetUserIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}